requirements.txt
# LangChain & LangGraph
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langgraph>=0.4.0

# OpenAI Python SDK (for audio transcription)
openai>=1.68.0

# FastAPI & Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# Environment variables
python-dotenv>=1.0.0

# Async support
aiofiles>=23.2.1

# Pydantic for data validation
pydantic>=2.0.0

# Audio processing (optional, for long audio files)
# pydub>=0.25.1
# audioop-lts>=0.2.0  # Python 3.13+ compatibility
